# Conflicts:
#	src/main.py

+++ src/model/ResultSet.py
@@ -4,10 +4,10 @@ from model.Experiment import EncodingParameters
 
 @dataclass
 class VideoResultMetrics:
-    Bitrate: int
-    PSNR: int
-    SSIM: int
-    VMAF: int
+    Bitrate: float
+    PSNR: float
+    SSIM: float
+    VMAF: float
 
 @dataclass
 class ResultSetItem:
